:hammer: Move file hash algorithm to commonMain for multiplatform reuse
